+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## 0.0.0 - Can we call this a release?
-* Package generation
-* Metadata filling
+## 0.0.0 - Autodetection from `dodo.py`
+* The package now parses the file `dodo.py` from the project directory
+* Tasks from `dodo.py` are registered as build targets in the `atom-build` package
+* The source code now contains tests to ensure correct parsing